requirements.txt
openai>=1.0.0
requests>=2.31.0
python-dotenv>=1.0.0
feedparser>=6.0.10
httpx>=0.24.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
pandas==2.1.4
numpy==1.24.3
urllib3==2.0.7
retry==0.9.2
ratelimit==2.2.1
python-dateutil==2.8.2
Jinja2==3.1.2 